requirements.txt
numpy~=1.19.2
attrs~=20.2.0
pytesseract~=0.3.6
opencv-python~=4.4.0.44
PyMuPDF